make cells and the arrow overlay render properly after dots conversion, jobify walk system

+++ Ported/larus/LabRat/Assets/Scripts/ECS/ArrowComponent.cs
@@ -7,5 +7,5 @@ public struct ArrowComponent : IComponentData
     public Direction Direction;
     public float2 Coordinates;
     public int PlayerId;
-    public int PlacementTick;
+    public float PlacementTick;
 }
